// src/app/services/api/model/nodeCountersQueryItem.ts
/**
 * TWCore Diagnostics Api
 * TWCore diagnostics api
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface NodeCountersQueryItem { 
    application?: string;
    countersId?: string;
    category?: string;
    name?: string;
    type?: NodeCountersQueryItem.TypeEnum;
    level?: NodeCountersQueryItem.LevelEnum;
    kind?: NodeCountersQueryItem.KindEnum;
    unit?: NodeCountersQueryItem.UnitEnum;
    typeOfValue?: string;
}
export namespace NodeCountersQueryItem {
    export type TypeEnum = 'Cumulative' | 'Average' | 'Current';
    export const TypeEnum = {
        Cumulative: 'Cumulative' as TypeEnum,
        Average: 'Average' as TypeEnum,
        Current: 'Current' as TypeEnum
    };
    export type LevelEnum = 'User' | 'Framework';
    export const LevelEnum = {
        User: 'User' as LevelEnum,
        Framework: 'Framework' as LevelEnum
    };
    export type KindEnum = 'Unknown' | 'Application' | 'DataAccess' | 'Messaging' | 'Cache' | 'RPC' | 'Bot';
    export const KindEnum = {
        Unknown: 'Unknown' as KindEnum,
        Application: 'Application' as KindEnum,
        DataAccess: 'DataAccess' as KindEnum,
        Messaging: 'Messaging' as KindEnum,
        Cache: 'Cache' as KindEnum,
        RPC: 'RPC' as KindEnum,
        Bot: 'Bot' as KindEnum
    };
    export type UnitEnum = 'Unknown' | 'Items' | 'Calls' | 'Milliseconds' | 'Seconds' | 'Minutes' | 'Hours' | 'Days' | 'Bytes' | 'MegaBytes' | 'Messages' | 'Exceptions';
    export const UnitEnum = {
        Unknown: 'Unknown' as UnitEnum,
        Items: 'Items' as UnitEnum,
        Calls: 'Calls' as UnitEnum,
        Milliseconds: 'Milliseconds' as UnitEnum,
        Seconds: 'Seconds' as UnitEnum,
        Minutes: 'Minutes' as UnitEnum,
        Hours: 'Hours' as UnitEnum,
        Days: 'Days' as UnitEnum,
        Bytes: 'Bytes' as UnitEnum,
        MegaBytes: 'MegaBytes' as UnitEnum,
        Messages: 'Messages' as UnitEnum,
        Exceptions: 'Exceptions' as UnitEnum
    };
}